Rename datapack top to factory
Renamed contents.datapack.top to contents.datapack.factory to make its role more explicitly clear
